Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@ int main(){
     scanf("%d %d %d",&a,&b,&c);
 
     float Average;
-    Average = (a + b + c)/3;
+    Average = (float) (a + b + c)/3;
 
     printf("Average: %.2f\n",Average);
     return 0;
